core/requirements.txt
aio_pika==9.6.2
apscheduler==3.11.1
fastapi==0.135.3
httpx==0.28.1
pandas==3.0.2
passlib==1.7.4
protobuf==7.34.1
pydantic==2.12.5
python-dotenv==1.2.2
python_jose==3.5.0
SQLAlchemy==2.0.49
